Sonatype plugin review changes
