+++ stage2/Publish.scala
@@ -15,8 +15,6 @@ trait Publish extends PackageJars{
 
   // ========== package ==========
 
-  // I guess pom should stay here, it's general for local and remote publish
-
   /** put additional xml that should go into the POM file in here */
   def pom: File = lib.pom(
     groupId = groupId,
@@ -41,6 +39,11 @@ trait Publish extends PackageJars{
 
   def publishLocal: Unit =
     lib.publishLocal( sourceFiles, `package` :+ pom, context.paths.mavenCache, releaseFolder )
+
+  def publishSnapshotLocal: Unit =
+    copy( context.copy(version = Some(version+"-SNAPSHOT")) ).publishLocal
+
+  def isSnapshot: Boolean = version.endsWith("-SNAPSHOT")
 
   override def copy(context: Context) = super.copy(context).asInstanceOf[Publish]
 }
